Fix m-values are doubles in dropped numerical limits export

+++ digiroad2-oracle/src/main/scala/fi/liikennevirasto/digiroad2/util/AssetDataImporter.scala
@@ -228,7 +228,7 @@ class AssetDataImporter {
     }
   }
 
-  def exportCsv(fileName: String, droppedLimits: List[(Long, Long, Int, Int, Int, Int, Boolean)]): Unit = {
+  def exportCsv(fileName: String, droppedLimits: List[(Long, Long, Double, Double, Int, Int, Boolean)]): Unit = {
     val headerLine = "mml_id; road_link_id; start_measure; end_measure; value \n"
     val data = droppedLimits.map { x =>
       s"""${x._1}; ${x._2}; ${x._3}; ${x._4}; ${x._5}"""
@@ -269,7 +269,7 @@ class AssetDataImporter {
            left join number_property_value s on s.asset_id = a.id
            where a.asset_type_id in (#${asset_name.keys.mkString(",")})
            and (valid_to is null or valid_to >= sysdate)
-         """.as[(Long, Long, Int, Int, Int, Int, Boolean)].list
+         """.as[(Long, Long, Double, Double, Int, Int, Boolean)].list
     }
     println("*** fetched all numerical limits from DB " + Seconds.secondsBetween(startTime, DateTime.now()).getSeconds)
